Move tests to new structure
